commit before changing mySingleton to store doubles in place of strings for all numeric values to try and preserve data precision in calculations.  Suspect that converting between strings and doubles/floats is flawed for number of decimal places it can use.

+++ VO2/VO2/calcsViewController.m
@@ -43,9 +43,9 @@
     // set up link to singleton
     mySingleton *singleton      = [mySingleton sharedSingleton];
     
-    float N2  = 0;
-    float HyN2= 0;
-    float O2  = 0;
+    double N2  = 0;
+    double HyN2= 0;
+    double O2  = 0;
     
     N2   = 100 - ([singleton.feo2 floatValue] + [singleton.feco2 floatValue]) ;
     HyN2 = N2 + (20.93 - [singleton.labO2 floatValue]);
